+++ cpp/thoth-ipc/include/thoth-ipc/abi_generated.hpp
@@ -120,6 +120,11 @@ inline constexpr const char* name_golden_ring = "__THOTH_SHM__QU_CONN__xchan__64
 #else
 inline constexpr const char* name_golden_ring = "__THOTH_SHM__QU_CONN__xchan__64__16";
 #endif
+#if defined(__APPLE__) && defined(__aarch64__)
+inline constexpr const char* name_golden_ring_posix = "/__THOTH_SHM___7d090bf7fa85c547";
+#else
+inline constexpr const char* name_golden_ring_posix = "/__THOTH_SHM__QU_CONN__xchan__64__16";
+#endif
 inline constexpr const char* name_golden_cc_id = "__THOTH_SHM__CA_CONN__";
 inline constexpr const char* name_golden_msg_id = "__THOTH_SHM__AC_CONN__xchan";
 inline constexpr const char* name_golden_liveness = "__THOTH_SHM__LV_CONN__xchan";