+++ rust/thoth-ipc/src/abi_generated.rs
@@ -120,6 +120,10 @@ pub const syncabi_stamp_payload_size_off: usize = 20;
 pub const name_golden_ring: &str = "__THOTH_SHM__QU_CONN__xchan__64__8";
 #[cfg(not(all(target_arch = "aarch64", target_vendor = "apple")))]
 pub const name_golden_ring: &str = "__THOTH_SHM__QU_CONN__xchan__64__16";
+#[cfg(all(target_arch = "aarch64", target_vendor = "apple"))]
+pub const name_golden_ring_posix: &str = "/__THOTH_SHM___7d090bf7fa85c547";
+#[cfg(not(all(target_arch = "aarch64", target_vendor = "apple")))]
+pub const name_golden_ring_posix: &str = "/__THOTH_SHM__QU_CONN__xchan__64__16";
 pub const name_golden_cc_id: &str = "__THOTH_SHM__CA_CONN__";
 pub const name_golden_msg_id: &str = "__THOTH_SHM__AC_CONN__xchan";
 pub const name_golden_liveness: &str = "__THOTH_SHM__LV_CONN__xchan";